==> advanced_tutorials/recommender-system/7_job_scheduler.ipynb <==
## <span style="color:#ff5f27">🗓️ Job Scheduling </span>


# !pip install hopsworks -q

from datetime import datetime, timezone

import hopsworks

project = hopsworks.login()

jobs_api = project.get_jobs_api()

## <span style="color:#ff5f27">🪝 Retrieve Jobs </span>


interactions_job = jobs_api.get_job('interactions_1_offline_fg_materialization')
interactions_job

transactions_job = jobs_api.get_job('transactions_1_offline_fg_materialization')
transactions_job

## <span style="color:#ff5f27">🏃🏻‍♂️ Run Job </span>


interactions_job_execution = interactions_job.run()
interactions_job_execution

transactions_job_execution = transactions_job.run()
transactions_job_execution

## <span style="color:#ff5f27">⏰ Schedule Job </span>


interactions_job.schedule(
    cron_expression="0 0 0 * * ?",  # Runs at midnight (00:00:00) every day
    start_time=datetime.now(tz=timezone.utc),
)
interactions_job.job_schedule.next_execution_date_time

transactions_job.schedule(
    cron_expression="0 0 0 * * ?",  # Runs at midnight (00:00:00) every day
    start_time=datetime.now(tz=timezone.utc),
)
transactions_job.job_schedule.next_execution_date_time

---
